+++ bin/dartion.dart
@@ -10,39 +10,39 @@ void main(List<String> arguments) async {
   final arg = arguments.isEmpty ? '--version' : arguments[0];
 
   if (arg == 'serve') {
-    var server = await DartIOServer.getInstance();
+    final server = await DartIOServer.getInstance();
     await server.start();
   } else if (arg == '--version' || arguments[0] == '-v') {
-    print('Dartion v$version');
+    stdout.write('Dartion v$version');
   } else if (arg == 'upgrade') {
     Process.runSync('pub', ['global', 'activate', 'dartion'], runInShell: true);
-    print('Upgrated!');
+    stdout.write('Upgrated!');
   } else if (arg == 'init') {
-    var dir = Directory(arguments.length > 1 ? arguments[1] : '.');
+    final dir = Directory(arguments.length > 1 ? arguments[1] : '.');
 
     if (!dir.existsSync()) {
       dir.createSync(recursive: true);
     }
 
-    print(dir.parent.path);
+    stdout.write(dir.parent.path);
     if (dir.listSync().isNotEmpty) {
-      print('Folder must be empty!');
+      stdout.write('Folder must be empty!');
       return;
     }
 
-    var db = File('${dir.path}/db.json');
+    final db = File('${dir.path}/db.json');
     db.createSync(recursive: true);
     db.writeAsStringSync(template.db);
 
-    var config = File('${dir.path}/config.yaml');
+    final config = File('${dir.path}/config.yaml');
     config.createSync(recursive: true);
     config.writeAsStringSync(template.config);
 
-    var index = File('${dir.path}/public/index.html');
+    final index = File('${dir.path}/public/index.html');
     index.createSync(recursive: true);
     index.writeAsStringSync(template.index);
-    print('Finished!');
-    print('Use dartio serve!');
+    stdout.write('Dartion initialization finished!');
+    stdout.write('You can now use the command: dartion serve');
   } else {
     exit(0);
   }