+++ lib/src/server/auth_service.dart
@@ -12,17 +12,33 @@ class AuthService {
     this.aud,
     this.scape,
   });
-  factory AuthService.formYaml(Map doc) {
+  factory AuthService.fromYaml(Map doc) {
     return AuthService(
       key: doc['key'],
       exp: doc['exp'],
-      aud: doc['aud'] == null ? null : (doc['aud'] as List).map<String>((e) => '$e').toList(),
-      scape: doc['scape'] == null ? [] : (doc['scape'] as List).map<String>((e) => '$e').toList(),
+      aud: doc['aud'] == null
+          ? null
+          : (doc['aud'] as List)
+              .map<String>(
+                (e) => '$e',
+              )
+              .toList(),
+      scape: doc['scape'] == null
+          ? []
+          : (doc['scape'] as List)
+              .map<String>(
+                (e) => '$e',
+              )
+              .toList(),
     );
   }
 
   String generateToken(int id) {
-    final claimSet = JwtClaim(subject: '$id', issuer: 'dartio', maxAge: Duration(seconds: exp));
+    final claimSet = JwtClaim(
+      subject: '$id',
+      issuer: 'dartio',
+      maxAge: Duration(seconds: exp),
+    );
 
     return issueJwtHS256(claimSet, key);
   }