+++ lib/src/server/dartio_server.dart
@@ -2,20 +2,19 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:path/path.dart';
 import 'package:shelf/shelf.dart';
 import 'package:shelf/shelf_io.dart' as shelf_io;
+import 'package:shelf_multipart/form_data.dart';
 import 'package:shelf_static/shelf_static.dart';
-import 'package:path/path.dart';
 import 'package:uuid/uuid.dart';
 
 import '../../dartion.dart';
-import '../config/config_model.dart';
-import 'package:shelf_multipart/form_data.dart';
 
 class DartIOServer {
   final Config config;
   late HttpServer _server;
-  var uuid = Uuid();
+  Uuid uuid = const Uuid();
 
   DartIOServer({required this.config});
 
@@ -27,16 +26,25 @@ class DartIOServer {
 
   Future start() async {
     await config.db.init();
-    var handler = const Pipeline().addMiddleware(logRequests()).addHandler(handleRequest);
+    final handler =
+        const Pipeline().addMiddleware(logRequests()).addHandler(handleRequest);
 
-    _server = await shelf_io.serve(handler, InternetAddress.loopbackIPv4, config.port);
-    print('Server ${config.name} started...');
-    print('Listening on localhost:${_server.port}');
+    _server = await shelf_io.serve(
+      handler,
+      InternetAddress.loopbackIPv4,
+      config.port,
+    );
+    stdout.write('Server ${config.name} started...');
+    stdout.write('Listening on localhost:${_server.port}');
   }
 
+  // request should be dynamic so the rule is ignored for this line
+  // ignore: type_annotate_public_apis
   bool checkFile(request) {
     if (request.uri.pathSegments.length >= 2) {
-      return request.uri.pathSegments[request.uri.pathSegments.length - 2] == 'file' && config.storage != null;
+      return request.uri.pathSegments[request.uri.pathSegments.length - 2] ==
+              'file' &&
+          config.storage != null;
     }
 
     return false;
@@ -44,12 +52,15 @@ class DartIOServer {
 
   FutureOr<Response> handleRequest(Request request) {
     try {
-      var mimeType = request.mimeType;
+      final mimeType = request.mimeType;
       final method = request.method.toUpperCase();
 
       if (method == 'GET') {
         if (request.url.pathSegments.last == config.statics) {
-          return createStaticHandler(config.statics, defaultDocument: 'index.html')(request);
+          return createStaticHandler(
+            config.statics,
+            defaultDocument: 'index.html',
+          )(request);
         } else if (request.url.pathSegments.last == config.storage?.folder) {
           return createStaticHandler(config.storage!.folder)(request);
         } else if (request.url.pathSegments.last == 'auth') {
@@ -61,7 +72,10 @@ class DartIOServer {
         return handleDelete(request);
       } else if (method == 'POST' && mimeType == 'application/json') {
         return handlePost(request);
-      } else if (method == 'POST' && mimeType == 'multipart/form-data' && request.url.pathSegments.last == 'storage' && config.storage != null) {
+      } else if (method == 'POST' &&
+          mimeType == 'multipart/form-data' &&
+          request.url.pathSegments.last == 'storage' &&
+          config.storage != null) {
         return handleUpload(request);
       } else if (method == 'PUT' && mimeType == 'application/json') {
         return handlePut(request);
@@ -81,7 +95,7 @@ class DartIOServer {
     }
   }
 
-  String get getSlash => Platform.isWindows ? '\\' : '/';
+  String get getSlash => Platform.isWindows ? r'\' : '/';
 
   Future<Response> handleUpload(Request request) async {
     if (!middlewareJwt(request)) {
@@ -91,41 +105,49 @@ class DartIOServer {
     if (!request.isMultipartForm) {
       return Response(401); // not a multipart request
     }
-    await for (final formData in request.multipartFormData.where((event) => event.name == config.storage?.name)) {
-      var dir = Directory(config.storage!.folder);
-      var name = "${dir.path}storage_${DateTime.now().millisecondsSinceEpoch}.${formData.filename?.split('.').last}";
+    await for (final formData in request.multipartFormData
+        .where((event) => event.name == config.storage?.name)) {
+      final dir = Directory(config.storage!.folder);
+      final name = '${dir.path}storage_${DateTime.now().millisecondsSinceEpoch}'
+          '.${formData.filename?.split('.').last}';
       final file = File(name);
       if (!file.existsSync()) {
         file.createSync(recursive: true);
       }
-      await file.writeAsBytes(await formData.part.readBytes(), mode: FileMode.writeOnly);
-      return Response.ok(jsonEncode({
-        'url': basename(file.path),
-      }));
+      await file.writeAsBytes(
+        await formData.part.readBytes(),
+        mode: FileMode.writeOnly,
+      );
+      return Response.ok(
+        jsonEncode({
+          'url': basename(file.path),
+        }),
+      );
     }
 
     return Response.internalServerError();
   }
 
   Future<Response> handleAuth(Request request) async {
-    var token = request.headers[HttpHeaders.authorizationHeader];
+    final token = request.headers[HttpHeaders.authorizationHeader];
     if (token == null) {
-      return Response.forbidden(jsonEncode({
-        'error': 'Not found token Basic',
-      }));
+      return Response.forbidden(
+        jsonEncode({
+          'error': 'Not found token Basic',
+        }),
+      );
     }
 
     try {
-      var credentials =
-        //O token é uma string, e ele estava sendo utilizado como token[0], o que retornaria apenas a letra B
+      //@Noslin22 fixes to credentials bug:
+      var credentials
           String.fromCharCodes(base64Decode(token.replaceFirst('Basic ', '')))
               .split(':');
       var users = await config.db.getAll('users');
       Map user = users.firstWhere((element) =>
           element['email'] == credentials[0] &&
           element['password'] == credentials[1]);
-      //Todo o codigo do newUser foi feito pois ele estava removendo o password do users, ou seja, se eu tentasse realizar o login novamente não funcionaria. 
-      //Provavelmente ele funcionaria em um servidor, mas acho que deve ser mehor assim, pois se trata de um plugin para testes de banco de dados
+     
       int index = user.keys.toList().indexOf("password");
       
       List keys = user.keys.toList();
@@ -137,6 +159,7 @@ class DartIOServer {
       return Response.ok(
         jsonEncode({
           'user': newUser,
+
           'token': config.auth?.generateToken(user['id']),
           'exp': config.auth?.exp
         }),
@@ -146,24 +169,26 @@ class DartIOServer {
       return Response.forbidden(jsonEncode({'error': 'Forbidden Access'}));
     }
   }
-
+  //end of @Noslin22 fixes 
+  
   bool middlewareJwt(Request request) {
     if (config.auth == null) {
       return true;
     }
 
-    if (request.url.pathSegments.isEmpty || config.auth?.scape?.contains(request.url.pathSegments[0]) == true) {
+    if (request.url.pathSegments.isEmpty ||
+        config.auth?.scape?.contains(request.url.pathSegments[0]) == true) {
       return true;
     }
 
-    var header = request.headers[HttpHeaders.authorizationHeader];
+    final header = request.headers[HttpHeaders.authorizationHeader];
     if (header == null) {
       return false;
     }
 
-    var token = header[0].replaceFirst('Bearer ', '');
+    final token = header[0].replaceFirst('Bearer ', '');
 
-    var valid = config.auth?.isValid(token, request.url.pathSegments[0]);
+    final valid = config.auth?.isValid(token, request.url.pathSegments[0]);
 
     if (valid != null) {
       return false;
@@ -174,7 +199,8 @@ class DartIOServer {
 
   Future<dynamic> getSegment(Request request) async {
     if (request.url.pathSegments.length > 1) {
-      return config.db.get(request.url.pathSegments.first, request.url.pathSegments[1]);
+      return config.db
+          .get(request.url.pathSegments.first, request.url.pathSegments[1]);
     } else {
       return config.db.getAll(request.url.pathSegments[0]);
     }
@@ -186,12 +212,15 @@ class DartIOServer {
     }
 
     try {
-      dynamic seg = await getSegment(request);
+      final dynamic seg = await getSegment(request);
 
       if (seg == null) {
         return Response.notFound(jsonEncode({'error': 'Not found'}));
       } else {
-        return Response.ok(jsonEncode(seg), headers: {'content-type': 'application/json'});
+        return Response.ok(
+          jsonEncode(seg),
+          headers: {'content-type': 'application/json'},
+        );
       }
     } catch (e) {
       return Response.notFound(jsonEncode({'error': 'Internal Error. $e'}));
@@ -203,10 +232,10 @@ class DartIOServer {
       return Response.forbidden(jsonEncode({'error': 'middlewareJwt'}));
     }
     try {
-      var content = await request.readAsString(); /*2*/
-      var data = jsonDecode(content) as Map;
+      final content = await request.readAsString(); /*2*/
+      final data = jsonDecode(content) as Map;
       final key = request.url.pathSegments[0];
-      dynamic seg = await config.db.getAll(request.url.pathSegments[0]);
+      final dynamic seg = await config.db.getAll(request.url.pathSegments[0]);
 
       if (seg == null) {
         return Response.notFound(jsonEncode({'error': 'Not found'}));
@@ -214,7 +243,10 @@ class DartIOServer {
         data['id'] = uuid.v1();
         seg.add(data);
         await config.db.save(key, seg);
-        return Response.ok(jsonEncode(data), headers: {'content-type': 'application/json'});
+        return Response.ok(
+          jsonEncode(data),
+          headers: {'content-type': 'application/json'},
+        );
       }
     } catch (e) {
       return Response.notFound(jsonEncode({'error': 'Internal Error. $e'}));
@@ -226,21 +258,26 @@ class DartIOServer {
       return Response.forbidden(jsonEncode({'error': 'middlewareJwt'}));
     }
     try {
-      var content = await request.readAsString(); /*2*/
-      var data = jsonDecode(content) as Map;
+      final content = await request.readAsString(); /*2*/
+      final data = jsonDecode(content) as Map;
       final key = request.url.pathSegments[0];
-      dynamic seg = await config.db.getAll(key);
+      final dynamic seg = await config.db.getAll(key);
 
       if (seg == null) {
         return Response.notFound(jsonEncode({'error': 'Not found'}));
       } else {
         data['id'] = request.url.pathSegments[1];
-        var position = (seg as List).indexWhere((element) => element['id'] == request.url.pathSegments[1]);
+        final position = (seg as List).indexWhere(
+          (element) => element['id'] == request.url.pathSegments[1],
+        );
         data.forEach((key, value) {
           seg[position][key] = value;
         });
         await config.db.save(key, seg);
-        return Response.ok(jsonEncode(data), headers: {'content-type': 'application/json'});
+        return Response.ok(
+          jsonEncode(data),
+          headers: {'content-type': 'application/json'},
+        );
       }
     } catch (e) {
       return Response.notFound(jsonEncode({'error': 'Internal Error. $e'}));
@@ -253,7 +290,7 @@ class DartIOServer {
     }
     try {
       final key = request.url.pathSegments[0];
-      dynamic seg = await config.db.getAll(key);
+      final dynamic seg = await config.db.getAll(key);
 
       if (seg == null) {
         return Response.notFound(jsonEncode({'error': 'Not found'}));
@@ -262,10 +299,15 @@ class DartIOServer {
           (element) => element['id'] == request.url.pathSegments[1],
         );
         await config.db.save(key, seg);
-        return Response.ok(jsonEncode({'data': 'ok!'}), headers: {'content-type': 'application/json'});
+        return Response.ok(
+          jsonEncode({'data': 'ok!'}),
+          headers: {'content-type': 'application/json'},
+        );
       }
     } catch (e) {
-      return Response.internalServerError(body: jsonEncode({'error': 'Internal Error'}));
+      return Response.internalServerError(
+        body: jsonEncode({'error': 'Internal Error'}),
+      );
     }
   }
 
@@ -275,27 +317,33 @@ class DartIOServer {
     }
 
     try {
-      var content = await request.readAsString(); /*2*/
-      var data = jsonDecode(content) as Map;
+      final content = await request.readAsString(); /*2*/
+      final data = jsonDecode(content) as Map;
       final key = request.url.pathSegments[0];
-
-      dynamic seg = await config.db.getAll(key);
+      final dynamic seg = await config.db.getAll(key);
 
       if (seg == null) {
         return Response.notFound(jsonEncode({'error': 'Not found'}));
       } else {
         data['id'] = int.parse(request.url.pathSegments[1]);
-        var position = (seg as List).indexWhere((element) => element['id'] == int.parse(request.url.pathSegments[1]));
+        final position = (seg as List).indexWhere(
+          (element) => element['id'] == int.parse(request.url.pathSegments[1]),
+        );
 
         data.forEach((key, value) {
           seg[position][key] = value;
         });
 
         await config.db.save(key, seg);
-        return Response.ok(jsonEncode(data), headers: {'content-type': 'application/json'});
+        return Response.ok(
+          jsonEncode(data),
+          headers: {'content-type': 'application/json'},
+        );
       }
     } catch (e) {
-      return Response.internalServerError(body: jsonEncode({'error': 'Internal Error'}));
+      return Response.internalServerError(
+        body: jsonEncode({'error': 'Internal Error'}),
+      );
     }
   }
 }