+++ CHANGELOG.md
@@ -1,8 +1,23 @@
-## 0.0.7 - 27 Apr 2020
+## 1.1.0 - Jan 2023
+[   ] Added Flutterando style readme
+- Added Flutterando_analysis as the new linter package
+- 77 Linter fixes (pending only "document all public members")
+[   ] Fixed Tests
+[   ] Added Noslin22's HandleAuth fix
+- Fixed issue #6 (String fix)
+- Changed methods wrongly indicated as "formYaml" to "fromYaml"
+- Updated package dependencies
+- Added example.md
+- Erased the 'Public' folder that had 3 index.html
+ 
+## 1.0.7 - 25 Nov 2021
+- Updated builds
+- Major changes
+- Several fixes
 
+## 0.0.7 - 27 Apr 2020
 - Added File upload (check documentation)
 
 ## 0.0.4 - 27 Apr 2020
-
 - Initial version